Move spectrum datasets to gammapy.datasets
