+++ gammapy/cube/psf_kernel.py
@@ -5,7 +5,6 @@ from astropy.coordinates import Angle
 from astropy.coordinates.angle_utilities import angular_separation
 from gammapy.irf import TablePSF
 from gammapy.maps import Map, WcsGeom
-from gammapy.modeling.models import PowerLawSpectralModel
 from gammapy.utils.gauss import Gauss2DPDF
 
 __all__ = ["PSFKernel"]
