+++ gammapy/datasets/__init__.py
@@ -1,1 +1,2 @@
 from .core import *
+from .dataset import *
