+++ gammapy/modeling/serialize.py
@@ -1,7 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Utilities to serialize models."""
 from gammapy.cube.fit import MapDataset
-from gammapy.spectrum import FluxPointsDataset, SpectrumDatasetOnOff
+from gammapy.spectrum import FluxPointsDataset
+from gammapy.datasets import SpectrumDatasetOnOff
 from .models import Registry, SkyDiffuseCube, SkyModel
 
 # TODO: move this elsewhere ?
