+++ gammapy/spectrum/__init__.py
@@ -1,7 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """1D spectrum analysis."""
 from .core import *
-from .dataset import *
 from .flux_point import *
 from .make import *
 from .phase import *
